Fix RuboCop offenses
Take out filling logic from the `Form` class.
